More precise composer map rotation (ticket #4632)

+++ src/app/composer/qgscomposermapwidget.cpp
@@ -163,7 +163,7 @@ void QgsComposerMapWidget::on_mScaleLineEdit_editingFinished()
   mComposerMap->endCommand();
 }
 
-void QgsComposerMapWidget::on_mRotationSpinBox_valueChanged( int value )
+void QgsComposerMapWidget::on_mRotationSpinBox_valueChanged( double value )
 {
   if ( !mComposerMap )
   {
